master post process proper

+++ scripts/post_process.py
@@ -14,7 +14,7 @@ def printThresh(l, threshold):
     threshCount = len(outliers)
     print("Values above " + str(threshold))
     print("     Number: " + str(threshCount))
-    percent = float('%.2f'%((threshCount/count) * 100))
+    percent = '%.2f'%((threshCount/count) * 100)
     print("     Ratio: " + str(threshCount) + "/" + str(count) + " | or " + percent + "%")
